+++ src/bin/todo-tree/app/commands/tags.rs
@@ -2,7 +2,7 @@ use super::load_config;
 use crate::app::cli;
 use crate::app::display::priority_to_color;
 use color_eyre::eyre::Result;
-use todo_tree::core::Priority;
+use todo_tree::core::TodoPriority;
 use todo_tree::core::tags::default_tag_names;
 
 pub fn run(args: cli::TagsArgs, global: &cli::GlobalOptions) -> Result<()> {
@@ -52,7 +52,7 @@ pub fn run(args: cli::TagsArgs, global: &cli::GlobalOptions) -> Result<()> {
             if global.no_color {
                 println!("  - {}", tag);
             } else {
-                let color = priority_to_color(Priority::from_tag(tag));
+                let color = priority_to_color(TodoPriority::from_tag(tag));
                 println!("  - {}", tag.color(color));
             }
         }